+++ src/Workspace.cpp
@@ -113,9 +113,15 @@ void WorkspaceFolder::indexFiles(const ClientConfiguration& config)
     if (isNullWorkspace())
         return;
 
+    size_t indexCount = 0;
+
     for (std::filesystem::recursive_directory_iterator next(rootUri.fsPath()), end; next != end; ++next)
     {
-        if (next->is_regular_file() && next->path().has_extension() && !isDefinitionFile(next->path(), config))
+        if (indexCount >= config.index.maxFiles)
+            break;
+
+        if (next->is_regular_file() && next->path().has_extension() && !isDefinitionFile(next->path(), config) &&
+            !isIgnoredFile(next->path(), config))
         {
             auto ext = next->path().extension();
             if (ext == ".lua" || ext == ".luau")
@@ -125,6 +131,7 @@ void WorkspaceFolder::indexFiles(const ClientConfiguration& config)
                 frontend.check(moduleName, Luau::FrontendOptions{/* retainFullTypeGraphs: */ true, /* forAutocomplete: */ true});
                 // TODO: do we need indexing for non-autocomplete?
                 // frontend.check(moduleName);
+                indexCount += 1;
             }
         }
     }